feat: :sparkles: Add active status validation to genre use cases

+++ src/modules/genres/application/use-cases/DeleteGenre.ts
@@ -1,3 +1,4 @@
+import { GenreIsActive } from "@genres/domain/exceptions/GenreIsActive";
 import { GenreWasNotFound } from "@genres/domain/exceptions/GenreWasNotFound";
 import type { GenreRepository } from "@genres/domain/repository/GenreRepository";
 import { GenreId } from "@genres/domain/value-objects/GenreId";
@@ -10,6 +11,8 @@ export class DeleteGenre {
     const genre = await this.repository.findById(genreId);
 
     if (!genre) throw new GenreWasNotFound("Genre was not found");
+
+    if (genre.isActive()) throw new GenreIsActive("Genre is active, cannot be deleted");
 
     await this.repository.delete(genreId);
   }
